(PC-33723)[API] chore: move EducationalBookingBaseResponse to core/educational/schemas.py
