+++ api/src/pcapi/core/educational/adage_backends/__init__.py
@@ -2,20 +2,17 @@ import datetime
 
 from pcapi import settings
 from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
+from pcapi.core.educational import schemas as educational_schemas
 from pcapi.core.educational.adage_backends import serialize
-from pcapi.core.educational.schemas import AdageCulturalPartner
-from pcapi.routes.adage.v1.serialization.prebooking import AdageReimbursementNotification
-from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingEdition
-from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingResponse
 from pcapi.utils.module_loading import import_string
 
 
-def notify_prebooking(data: EducationalBookingResponse) -> None:
+def notify_prebooking(data: educational_schemas.EducationalBookingResponse) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_prebooking(data=data)
 
 
-def notify_offer_or_stock_edition(data: EducationalBookingEdition) -> None:
+def notify_offer_or_stock_edition(data: educational_schemas.EducationalBookingEdition) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_offer_or_stock_edition(data=data)
 
@@ -26,7 +23,7 @@ def get_adage_offerer(siren: str) -> list[AdageVenue]:
     return result
 
 
-def notify_booking_cancellation_by_offerer(data: EducationalBookingResponse) -> None:
+def notify_booking_cancellation_by_offerer(data: educational_schemas.EducationalBookingResponse) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_booking_cancellation_by_offerer(data=data)
 
@@ -42,7 +39,7 @@ def notify_institution_association(data: serialize.AdageCollectiveOffer) -> None
     backend().notify_institution_association(data=data)
 
 
-def get_cultural_partner(siret: str) -> AdageCulturalPartner:
+def get_cultural_partner(siret: str) -> educational_schemas.AdageCulturalPartner:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().get_cultural_partner(siret)
     return result
@@ -60,7 +57,7 @@ def get_adage_educational_redactor_from_uai(uai: str) -> list[dict[str, str]]:
     return result
 
 
-def notify_reimburse_collective_booking(data: AdageReimbursementNotification) -> None:
+def notify_reimburse_collective_booking(data: educational_schemas.AdageReimbursementNotification) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().notify_reimburse_collective_booking(data)
     return result
