+++ api/src/pcapi/core/educational/adage_backends/base.py
@@ -3,24 +3,23 @@ import datetime
 from pcapi import settings
 from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
 from pcapi.core.educational.adage_backends import serialize
-from pcapi.core.educational.schemas import AdageCulturalPartner
-from pcapi.routes.adage.v1.serialization import prebooking
+import pcapi.core.educational.schemas as educational_schemas
 
 
 class AdageClient:
     def __init__(self) -> None:
         self.base_url = settings.ADAGE_API_URL
 
-    def notify_prebooking(self, data: prebooking.EducationalBookingResponse) -> None:
+    def notify_prebooking(self, data: educational_schemas.EducationalBookingResponse) -> None:
         raise NotImplementedError()
 
-    def notify_offer_or_stock_edition(self, data: prebooking.EducationalBookingEdition) -> None:
+    def notify_offer_or_stock_edition(self, data: educational_schemas.EducationalBookingEdition) -> None:
         raise NotImplementedError()
 
     def get_adage_offerer(self, siren: str) -> list[AdageVenue]:
         raise NotImplementedError()
 
-    def notify_booking_cancellation_by_offerer(self, data: prebooking.EducationalBookingResponse) -> None:
+    def notify_booking_cancellation_by_offerer(self, data: educational_schemas.EducationalBookingResponse) -> None:
         raise NotImplementedError()
 
     def get_cultural_partners(
@@ -31,7 +30,7 @@ class AdageClient:
     def notify_institution_association(self, data: serialize.AdageCollectiveOffer) -> None:
         raise NotImplementedError()
 
-    def get_cultural_partner(self, siret: str) -> AdageCulturalPartner:
+    def get_cultural_partner(self, siret: str) -> educational_schemas.AdageCulturalPartner:
         raise NotImplementedError()
 
     def get_adage_educational_institutions(self, ansco: str) -> list[serialize.AdageEducationalInstitution]:
@@ -40,7 +39,7 @@ class AdageClient:
     def get_adage_educational_redactor_from_uai(self, uai: str) -> list[dict[str, str]]:
         raise NotImplementedError()
 
-    def notify_reimburse_collective_booking(self, data: prebooking.AdageReimbursementNotification) -> None:
+    def notify_reimburse_collective_booking(self, data: educational_schemas.AdageReimbursementNotification) -> None:
         raise NotImplementedError()
 
     def notify_redactor_when_collective_request_is_made(self, data: serialize.AdageCollectiveRequest) -> None:
