+++ api/src/pcapi/routes/adage/v1/serialization/educational_institution.py
@@ -1,14 +1,13 @@
 from pydantic.v1.fields import Field
 
-from pcapi.core.educational.schemas import AdageBaseResponseModel
-from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingResponse
+import pcapi.core.educational.schemas as educational_schemas
 from pcapi.serialization.utils import to_camel
 
 
-class EducationalInstitutionResponse(AdageBaseResponseModel):
+class EducationalInstitutionResponse(educational_schemas.AdageBaseResponseModel):
     credit: int = Field(description="Total credit granted to the educational institution")
     isFinal: bool = Field(description="Flag to know if the credit has been approved and is now final")
-    prebookings: list[EducationalBookingResponse]
+    prebookings: list[educational_schemas.EducationalBookingResponse]
 
     class Config:
         title = "School response model"
