+++ api/tests/routes/pro/patch_collective_offer_test.py
@@ -10,6 +10,7 @@ import pcapi.core.educational.factories as educational_factories
 from pcapi.core.educational.models import CollectiveBookingStatus
 from pcapi.core.educational.models import CollectiveOffer
 from pcapi.core.educational.models import StudentLevels
+from pcapi.core.educational.schemas import EducationalBookingEdition
 import pcapi.core.educational.testing as educational_testing
 import pcapi.core.finance.factories as finance_factories
 import pcapi.core.finance.models as finance_models
@@ -19,7 +20,6 @@ import pcapi.core.providers.factories as providers_factories
 from pcapi.core.testing import override_features
 import pcapi.core.users.factories as users_factories
 from pcapi.models import db
-from pcapi.routes.adage.v1.serialization.prebooking import EducationalBookingEdition
 from pcapi.routes.adage.v1.serialization.prebooking import serialize_collective_booking
